+++ src/App.js
@@ -1,15 +1,15 @@
 import React from "react";
 import "./App.css";
 import MeditationComponent from "./MeditationComponent";
-import TrainInfoComponent from "./TrainInfoComponent";
+import TrainInfoComponent from "./components/TrainInfoComponent/TrainInfoComponent";
 
 function App() {
   return (
-  <div>
-  <TrainInfoComponent />;
-  <MeditationComponent />;
-  </div>
-  )
+    <div>
+      <TrainInfoComponent />
+      <MeditationComponent />
+    </div>
+  );
 }
 
 export default App;
