--- src/HardCodeTrainInfo.js
@@ -1,16 +1,0 @@
-import React from "react";
-
-function HardCodeTrainInfo () {
-    return (
-    <div className = "info-box">
-        <h1 className = "first-line">TRAVEL INFORMATION</h1>
-        <div id="info-wrapper">
-            <h3 className = "delays-line">DELAYS</h3>
-            <h1 className = "station-line">LUND C</h1>
-            <h1 className = "time-line">5 MIN</h1>
-        </div>
-    </div>
-    )
-}
-
-export default HardCodeTrainInfo;
